requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-jose[cryptography]
passlib[bcrypt]
authlib
httpx
python-multipart
itsdangerous
pytest
pytest-asyncio
beautifulsoup4
yfinance
curl_cffi
PyYAML
psycopg2-binary
pymysql
